Fix a bug related to fileparts behaviour (Windows)
In calls like (with GNU Octave for Windows 5.2.0):

[directory, name, ext] = fileparts ('C:\')

fileparts is eating the backslash and keeps only "C:" in 'directory'
(and empty strings in 'name' and 'ext').

More investigation would be needed:
- Does Matlab do the same?
- Do older Octave versions do the same?

This behaviour implied that a loop in 'detect_test_cases' did not detect
that it had reached the file system root and continued with a call like:

[directory, name] = fileparts('C:')

which gives 'directory' empty and "C:" in 'name'. 'directory' empty
causes an error when evaluating 'directory(end)' in the loop condition.

The fix robustifies the root detection with a 'directory(end) = ':''
test in the loop condition (only under Windows).